Refactoring sunburst code into PlotlySunburstPlot
